# server/src/swagger/routes.yaml
swagger: '3.0.0'
info:
  title: Kids Activity Tracker API
  version: '1.0.0'
  description: |
    Comprehensive API documentation for the Kids Activity Tracker application.
    
    ## API Overview
    
    ### Core Concepts:
    - **Categories**: 5 age-based groupings (Early Years: Parent Participation, Early Years: On My Own, School Age, Youth, All Ages & Family)
    - **Activity Types**: 22 subject-based groupings (Swimming & Aquatics, Team Sports, etc.)
    - **Global Filters**: hideClosedActivities, hideFullActivities (apply to all endpoints except Favorites)
    - **Pagination**: 50 items per page with offset-based pagination
    
    ### Navigation Flows:
    1. **Browse by Location**: Dashboard → Cities → Venues → Activities
    2. **Browse by Category**: Dashboard → Categories → Category Detail → Activities  
    3. **Browse by Activity Type**: Dashboard → Activity Types → Activity Type Detail → Activities
    4. **Recommendations**: Dashboard → Recommended Activities (user-specific)
    
    ### React Native Components Usage:
    - **DashboardScreen**: Uses categories, activity types, locations APIs
    - **CategoryDetailScreen**: Uses `/categories/{id}/activities`
    - **AllCategoriesScreen**: Uses `/categories`
    - **LocationBrowseScreen**: Uses `/locations/cities`, `/locations/{cityId}/venues`
    - **ActivityDetailScreen**: Uses `/activities/{id}`, favorites APIs
    - **SearchScreen**: Uses `/activities/search` with extensive filtering
paths:
...
  /api/v1/locations/cities:
    get:
      tags: [Locations]
      summary: Get all cities with venue and activity counts
      description: |
        Returns all cities with their venue and activity counts.
        Used by: LocationBrowseScreen for city selection.
        Navigation flow: Dashboard → Browse by Location → Cities (this endpoint) → Venues → Activities
      parameters:
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities from counts
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities from counts
        - in: query
          name: includeActivityCounts
          schema:
            type: boolean
            default: true
          description: Whether to include activity counts (can be disabled for performance)
      responses:
        200:
          description: List of cities with counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cities:
                    type: array
                    items:
                      $ref: '#/components/schemas/City'
...
  /api/v1/cities/{city}/locations:
    get:
      tags: [Cities]
      summary: Get locations in city
      description: |
        Get all locations/venues in a specific city.
        Used by: City location screens
      parameters:
        - in: path
          name: city
          required: true
          schema:
            type: string
      responses:
        200:
          description: City locations
  /api/v1/cities/{city}/activities:
    get:
      tags: [Cities]
      summary: Get activities in city
      description: |
        Get all activities in a specific city with pagination.
        Used by: City activities screens
      parameters:
        - in: path
          name: city
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: City activities

  # ==================== REFERENCE ENDPOINTS ====================
...
components:
  schemas:
...
    City:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        province:
          type: string
        country:
          type: string
        venueCount:
          type: integer
        activityCount:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...